+++ examples/tanstack-start/app/components/test.tsx
@@ -1,0 +1,14 @@
+import { type ReactNode } from 'react'
+
+interface TestProps {
+  children?: ReactNode
+}
+
+export function Test({ children }: TestProps) {
+  return (
+    <div>
+        <p>Test Component</p>
+        {children}
+    </div>
+  )
+}